training/lora-trainer/requirements.txt
torch==2.1.0
torchvision
transformers==4.35.0
diffusers==0.24.0
accelerate==0.24.0
peft==0.6.0
bitsandbytes==0.41.0
safetensors
xformers
google-cloud-storage
Pillow
numpy
tqdm
datasets
tensorboard